feat : 직업 / 역량 데이터 추가 (#12)
* feat : 데이터 형식 수정

* feat : 직업 / 역량 데이터 추가

+++ chat-da/src/main/java/site/chatda/domain/job/entity/JobSkill.java
@@ -25,6 +25,6 @@ public class JobSkill {
 
     @ManyToOne(fetch = LAZY, optional = false)
     @MapsId("skillId")
-    @JoinColumn(name = "skill_id", columnDefinition = "TINYINT")
+    @JoinColumn(name = "skill_id", columnDefinition = "MEDIUMINT")
     private Skill skill;
 }
